DDCNL-3749: All review comments completed besides scalatestplus
