+++ build.sbt
@@ -7,6 +7,8 @@ val appName: String = "tai"
 
 lazy val playSettings: Seq[Setting[_]] = Seq(routesImport ++= Seq("uk.gov.hmrc.tai.binders._", "uk.gov.hmrc.domain._"))
 
+val silencerVersion = "1.7.3"
+
 lazy val microservice = Project(appName, file("."))
   .enablePlugins(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtDistributablesPlugin)
   .settings(playSettings ++ scoverageSettings: _*)
@@ -17,7 +19,7 @@ lazy val microservice = Project(appName, file("."))
     evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
     routesGenerator := InjectedRoutesGenerator,
     PlayKeys.playDefaultPort := 9331,
-    scalaVersion := "2.12.12"
+    scalaVersion := "2.12.13"
   )
   .settings(inConfig(TemplateTest)(Defaults.testSettings): _*)
   .configs(IntegrationTest)
@@ -31,6 +33,14 @@ lazy val microservice = Project(appName, file("."))
       "uk.gov.hmrc.tai.model.domain._")
   )
   .settings(majorVersion := 0)
+  .settings(
+    scalacOptions += "-P:silencer:pathFilters=views;routes",
+    libraryDependencies ++= Seq(
+      compilerPlugin(
+        "com.github.ghik" % "silencer-plugin" % silencerVersion cross CrossVersion.full),
+      "com.github.ghik" % "silencer-lib" % silencerVersion % Provided cross CrossVersion.full
+    )
+  )
 
 lazy val scoverageSettings = {
   import scoverage.ScoverageKeys
