+++ test/uk/gov/hmrc/tai/connectors/RtiConnectorSpec.scala
@@ -62,7 +62,10 @@ class RtiConnectorSpec extends ConnectorBaseSpec {
         .withHeader("Authorization", equalTo("Bearer Local"))
         .withHeader("Gov-Uk-Originator-Id", equalTo(desOriginatorId))
         .withHeader(HeaderNames.xSessionId, equalTo(sessionId))
-        .withHeader(HeaderNames.xRequestId, equalTo(requestId)))
+        .withHeader(HeaderNames.xRequestId, equalTo(requestId))
+        .withHeader(
+          "CorrelationId",
+          matching("[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}")))
 
   "RtiConnector" when {
